+++ src/pages/Dashboard/SuperAdmin/HandleOrganizers/HandleOrganizers.jsx
@@ -21,27 +21,33 @@ const HandleOrganizers = () => {
   const navigateOrganizer = (orgId) => {
     navigate(`/updateOrganizer/${orgId}`);
   };
+  const navigateOrgPackages = (oId) => {
+    navigate(`/organizerDetails/${oId}`);
+  };
 
   // delete organization
   const handleDelete = (id) => {
     const url = `https://event-api.nurhossainfarid.com/v1/organization/${id}`;
-    fetch(url, {
-      method: "DELETE",
-      headers: {
-        "Content-Type": "application/json",
-      },
-    })
-      .then((res) => res.json())
-      .then((result) => {
-        if (result.status.toLowerCase() === "success") {
-          toast.success("Organizer delete successfully");
-        } else {
-          toast.error("Organizer could not delete successfully");
-        }
-      });
+    const decision = window.confirm("Will you remove this Organizer?");
+    if (decision) {
+      fetch(url, {
+        method: "DELETE",
+        headers: {
+          "Content-Type": "application/json",
+        },
+      })
+        .then((res) => res.json())
+        .then((result) => {
+          if (result.status.toLowerCase() === "success") {
+            toast.success("Organizer delete successfully");
+          } else {
+            toast.error("Organizer could not delete successfully");
+          }
+        });
+    }
   };
   return (
-    <div className="mx-1 md:mx-0 mt-5 md:mt-10">
+    <div className="mx-1 md:mx-0 mt-5 md:mt-10 pb-10">
       <div className="flex justify-between">
         <h1 className="text-xl md:text-3xl text-center mb-10 dashboard-header">
           Handle Organizers
@@ -54,6 +60,8 @@ const HandleOrganizers = () => {
           Add Organizers
         </Link>
       </div>
+      <p className='text-lg'>Click the <span className="text-purple-600 font-semibold">"Details" </span> 
+           Button to Know about organizer details, also update or delete a organizer</p>
       <table className="table w-full">
         <thead>
           <tr>
@@ -68,28 +76,78 @@ const HandleOrganizers = () => {
         <tbody>
           {organizersData.map((org) => (
             <tr>
-              <td>{org?.name}</td>
+              <td>{(org?.name).substring(0, 30)}...</td>
               <td>{org.email}</td>
               <td>{org.contactNumber}</td>
-              <td>{org.address}</td>
+              <td>{(org.address).substring(0, 15)}..</td>
               <td>{org.status}</td>
               <td className="flex flex-col gap-2">
-                <button className="btn btn-sm hover:btn-secondary">
-                  Details
-                </button>
-                <button
-                  className="btn btn-sm hover:btn-secondary"
-                  onClick={() => navigateOrganizer(org._id)}
-                >
-                  Update
-                </button>
-                <button
-                  className="btn btn-sm hover:btn-secondary"
-                  onClick={() => handleDelete(org._id)}
-                >
-                  Delete
-                </button>
-              </td>
+                    {/* The button to open modal */}
+                  <label htmlFor={org?._id} className="btn btn-xs 
+                    hover:btn-secondary">Details</label>
+
+                    {/* Put this part before </body> tag */}
+                    <input type="checkbox" id={org?._id} className="modal-toggle" />
+                    <div className="modal">
+                      <div className="modal-box w-7/12 max-w-5xl">
+                        <label htmlFor={org?._id} className="btn btn-sm btn-circle absolute right-2 
+                      top-2">✕</label>
+                        <div className="grid grid-cols-1 gap-5">
+                          <p className="text-lg"><span className="font-semibold text-sm">Name : </span> 
+                           {org?.name} 
+                           </p>
+                          <p className="text-lg"><span className="font-semibold text-sm">Email : </span> 
+                           {org?.email} 
+                           </p>
+                          <p className="text-lg"><span className="font-semibold text-sm">Phone : </span> 
+                           {org?.contactNumber} 
+                           </p>
+                          <p className="text-lg"><span className="font-semibold text-sm">Address : </span> 
+                           {org?.address}  
+                           </p>
+                          <p className="text-lg"><span className="font-semibold text-sm">Status : </span> 
+                           {org?.status} 
+                           </p>
+                          <p className="text-lg"><span className="font-semibold text-sm">Image : </span> 
+                           {org?.image}  
+                           </p>
+                          <p className="text-lg"><span className="font-semibold text-sm">Website : </span> 
+                           {org?.website}  
+                           </p>
+                          <p className="text-lg"><span className="font-semibold text-sm">Facebook : </span> 
+                           {org?.facebook}  
+                           </p>
+                          <p className="text-lg"><span className="font-semibold text-sm">Whatsapp : </span> 
+                           {org?.whatsapp}  
+                           </p>
+                          <p className="text-lg"><span className="font-semibold text-sm">Instagram : </span> 
+                           {org?.instagram}  
+                           </p>
+                        </div>
+                    <div className="modal-action">
+                          <button
+                            className="btn btn-sm hover:btn-secondary"
+                            onClick={() => navigateOrgPackages(org?._id)}
+                          >
+                            See Packages
+                          </button>
+                          <button
+                            className="btn btn-sm hover:btn-secondary"
+                            onClick={() => navigateOrganizer(org._id)}
+                          >
+                            Update
+                          </button>
+                      <button
+                          htmlFor={org?._id}
+                            className="btn btn-sm hover:btn-secondary"
+                            onClick={() => handleDelete(org._id)}
+                          >
+                            Delete
+                          </button>
+                        </div>
+                      </div>
+                    </div>
+                </td>
             </tr>
           ))}
         </tbody>
