+++ src/pages/Dashboard/SuperAdmin/HandleUser/HandleUser.jsx
@@ -22,20 +22,23 @@ const UserHistory = () => {
   // delete user
   const handleDelete = (id) => {
     const url = `https://event-api.nurhossainfarid.com/v1/user/${id}`;
-    fetch(url, {
-      method: "DELETE",
-      headers: {
-        "Content-Type": "application/json",
-      },
-    })
-      .then((res) => res.json())
-      .then((result) => {
-        if (result.status === "success") {
-          toast.success("User deleted successfully");
-        } else {
-          toast.error("User could not be deleted successfully");
-        }
-      });
+    const decision = window.confirm("Will you remove this user?");
+    if (decision) {  
+      fetch(url, {
+        method: "DELETE",
+        headers: {
+          "Content-Type": "application/json",
+        },
+      })
+        .then((res) => res.json())
+        .then((result) => {
+          if (result.status.toLowerCase() === "success") {
+            toast.success("User deleted successfully");
+          } else {
+            toast.error("User could not be deleted successfully");
+          }
+        });
+    }
   };
 
   // make super admin
@@ -43,19 +46,27 @@ const UserHistory = () => {
     const updateRole = {
       role: "super admin",
     };
-    fetch(`https://event-api.nurhossainfarid.com/v1/user/${id}`, {
-      method: "PATCH",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify(updateRole),
-    })
-      .then((res) => {
-        res.json();
+    const decision = window.confirm("Will you update this user role?");
+    if (decision) {  
+      fetch(`https://event-api.nurhossainfarid.com/v1/user/${id}`, {
+        method: "PATCH",
+        headers: {
+          "Content-Type": "application/json",
+        },
+        body: JSON.stringify(updateRole),
       })
-      .then((data) => {
-        console.log(data);
-      });
+        .then((res) => {
+          res.json();
+        })
+        .then((data) => {
+          console.log(data);
+          // if (result.status.toLowerCase() === "success") {
+          //   toast.success("Organizer delete successfully");
+          // } else {
+          //   toast.error("Organizer could not delete successfully");
+          // }
+        });
+    }
   };
 
   // make organizer admin
@@ -63,38 +74,38 @@ const UserHistory = () => {
     const updateRole = {
       role: "organizerAdmin",
     };
-    fetch(`https://event-api.nurhossainfarid.com/v1/user/${id}`, {
-      method: "PATCH",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify(updateRole),
-    })
-      .then((res) => {
-        res.json();
+    const decision = window.confirm("Will you update this user role?");
+    if (decision) {   
+      fetch(`https://event-api.nurhossainfarid.com/v1/user/${id}`, {
+        method: "PATCH",
+        headers: {
+          "Content-Type": "application/json",
+        },
+        body: JSON.stringify(updateRole),
       })
-      .then((data) => {
-        console.log(data);
-      });
+        .then((res) => {
+          res.json();
+        })
+        .then((data) => {
+          console.log(data);
+        });
+    }
   };
   return (
     <div className="mt-5">
-      <h1 className="dashboard-header text-xl md:text-3xl text-center animate-bounce">
-        Handle User
-      </h1>
       <div className="overflow-x-auto mt-5">
+        <p className='text-lg'>Click the <span className="text-purple-600 font-semibold">"Details" </span> 
+           Button to Know about user details, also update or delete a user and change user role.</p>
         <table className="table w-full">
           <thead>
             <tr>
               <th>First Name</th>
               <th>Last Name</th>
               <th>Email</th>
-              <th>Address</th>
               <th>Contact Number</th>
               <th>Role</th>
               <th>Status</th>
               <th>Action</th>
-              <th>Make Admin</th>
             </tr>
           </thead>
           <tbody>
@@ -103,53 +114,95 @@ const UserHistory = () => {
                 <td>{user.firstName}</td>
                 <td>{user.lastName}</td>
                 <td>{user.email}</td>
-                <td>{user.address}</td>
                 <td>{user.contactNumber}</td>
                 <td>{user.role}</td>
                 <td>{user.status}</td>
                 <td className="flex flex-col gap-2">
-                  <button
-                    className="btn btn-xs hover:btn-secondary"
-                    onClick={() => navigateUser(user?._id)}
-                  >
-                    Update
-                  </button>
-                  <button
-                    onClick={() => handleDelete(user._id)}
-                    className="btn btn-xs hover:btn-secondary"
-                  >
-                    Delete
-                  </button>
+                    {/* The button to open modal */}
+                  <label htmlFor={user?._id} className="btn btn-xs 
+                    hover:btn-secondary">Details</label>
+
+                    {/* Put this part before </body> tag */}
+                    <input type="checkbox" id={user?._id} className="modal-toggle" />
+                    <div className="modal">
+                      <div className="modal-box w-6/12 max-w-5xl">
+                        <label htmlFor={user?._id} className="btn btn-sm btn-circle absolute right-2 
+                      top-2">✕</label>
+                        <div className="grid grid-cols-2 gap-5">
+                          <p className="text-lg"><span className="font-semibold text-sm">First Name : </span> 
+                           {user?.firstName} 
+                           </p>
+                          <p className="text-lg"><span className="font-semibold text-sm">Last Name : </span> 
+                           {user?.lastName} 
+                           </p>
+                          <p className="text-lg"><span className="font-semibold text-sm">Email : </span> 
+                           {user?.email} 
+                           </p>
+                          <p className="text-lg"><span className="font-semibold text-sm">Phone : </span> 
+                           {user?.contactNumber}  
+                           </p>
+                          <p className="text-lg"><span className="font-semibold text-sm">Address : </span> 
+                           {user?.address} 
+                           </p>
+                          <p className="text-lg"><span className="font-semibold text-sm">Role : </span> 
+                           {user?.role} 
+                           </p>
+                          <p className="text-lg"><span className="font-semibold text-sm">Status : </span> 
+                           {user?.status}  
+                           </p>
+                      </div>
+                      <div className="flex justify-evenly mt-10">
+                        <div className="">
+                          <p className="text-center my-2">Make Admin</p>
+                            {user.role === "user" && (
+                              <div className="flex gap-3">
+                                <button
+                                  onClick={() => makeSuperAdmin(user._id)}
+                                  className="btn btn-xs hover:btn-secondary "
+                                >
+                                  super Admin
+                                </button>
+                                <button
+                                  className="btn btn-xs hover:btn-secondary"
+                                  onClick={() => makeOrganizerAdmin(user._id)}
+                                >
+                                  Organizer Admin
+                                </button>
+                              </div>
+                            )}
+                            {user.role === "organizeradmin" && (
+                              <div className="flex flex-col gap-2">
+                                <button
+                                  onClick={() => makeSuperAdmin(user._id)}
+                                  className="btn btn-xs hover:btn-secondary "
+                                >
+                                  super Admin
+                                </button>
+                              </div>
+                            )}
+                          </div>
+                        <div className="">
+                          <p className="text-center my-2">Action</p>
+                          <div className="flex gap-3">
+                            <button
+                              className="btn btn-xs hover:btn-secondary"
+                              onClick={() => navigateUser(user?._id)}
+                            >
+                              Update
+                            </button>
+                            <button
+                              onClick={() => handleDelete(user._id)}
+                              className="btn btn-xs hover:btn-secondary"
+                            >
+                              Delete
+                            </button>
+                          </div>
+                        </div>
+                      </div>
+                    </div>
+                  </div>
                 </td>
                 <td className="">
-                  <div>
-                    {user.role === "user" && (
-                      <div className="flex flex-col gap-2">
-                        <button
-                          onClick={() => makeSuperAdmin(user._id)}
-                          className="btn btn-xs hover:btn-secondary "
-                        >
-                          super Admin
-                        </button>
-                        <button
-                          className="btn btn-xs hover:btn-secondary"
-                          onClick={() => makeOrganizerAdmin(user._id)}
-                        >
-                          Organizer Admin
-                        </button>
-                      </div>
-                    )}
-                    {user.role === "organizeradmin" && (
-                      <div className="flex flex-col gap-2">
-                        <button
-                          onClick={() => makeSuperAdmin(user._id)}
-                          className="btn btn-xs hover:btn-secondary "
-                        >
-                          super Admin
-                        </button>
-                      </div>
-                    )}
-                  </div>
                 </td>
               </tr>
             ))}
